EdgeModResult
many edge modifications result in either a single face being created, a single face being destroyed, or faces being preserved. Thus we have added EdgeModResult to return which state is true. The exception is insertAfter, which does not return info on the root removal op that is performed before insertion, if this info is relevant, use a moveroot THEN insertAfter to aquire both.

Also added merge test cases and edited merge to use this new structure