Integrated Region
This allows for traversal between regions, easier tracking of resources, and better management of region internals. This comes with a few repercussions, face ordering is now dependent on order added, invalidating many tests that relied on the old, looser, ordering. This can be fixed easily. It also means that many operations that sat in region are now external to the system and must be passed a region to reference.